+++ src/lib/cloudinary.tsx
@@ -1,0 +1,7 @@
+import {v2} from 'cloudinary'
+
+export const cloudinary = v2.config({
+    cloud_name: '',
+    api_key: '',
+    api_secret: ''
+})